# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于Gin框架和Gorm二次封装的Go语言Web开发框架",
        "title": "inis CMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "兔子 or racns",
            "url": "https://github.com/racns/inis",
            "email": "97783391@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "paths": {
...
        "/api/tags/{method}": {
            "get": {
                "description": "根据不同方法获取标签相关数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "获取标签数据",
                "parameters": [
                    {
                        "enum": [
                            "one",
                            "all",
                            "sum",
                            "min",
                            "max",
                            "rand",
                            "count",
                            "column"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询条件",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字段过滤",
                        "name": "field",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新标签数据或恢复已删除的标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "更新/恢复标签",
                "parameters": [
                    {
                        "enum": [
                            "update",
                            "restore"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "标签描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "标签头像",
                        "name": "avatar",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "JSON数据",
                        "name": "json",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "文本数据",
                        "name": "text",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "标签ID列表（用于恢复）",
                        "name": "ids",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新标签或保存标签数据（包含创建和更新）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "创建/保存标签",
                "parameters": [
                    {
                        "enum": [
                            "save",
                            "create"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标签描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "标签头像",
                        "name": "avatar",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "JSON数据",
                        "name": "json",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "文本数据",
                        "name": "text",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应，包含标签ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "软删除、真实删除标签或清空回收站",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "删除标签",
                "parameters": [
                    {
                        "enum": [
                            "remove",
                            "delete",
                            "clear"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标签ID列表（逗号分隔）",
                        "name": "ids",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}